backend/requirements.txt
fastapi>=0.88.0,<0.96.0
pydantic>=1.10.0,<1.11.0
uvicorn>=0.20.0,<0.22.0
sqlalchemy>=2.0.0,<2.1.0
snowflake-sqlalchemy>=1.4.0,<1.7.0
boto3>=1.26.0,<1.27.0
mangum>=0.17.0,<0.18.0
python-dotenv>=1.0.0,<=1.0.1